+++ crates/cairo-profiler/tests/e2e.rs
@@ -957,9 +957,9 @@ fn tree_more_nested_calls_than_triggers_missing_call_contract() {
         .failure()
         .stderr_eq(indoc!(
             r#"
-            [ERROR] There are no syscalls left in the program trace, but at least one unhandled call in trace file CallEntryPoint { class_hash: Some(ClassHash(0x117)), entry_point_type: External, entry_point_selector: EntryPointSelector(0x17340c6779204ea2a91c87d1c2226a3aebda65c64da3672a36893c4330ea27b), contract_address: ContractAddress(0x1724987234973219347210837402), call_type: Call, contract_name: Some("SNFORGE_TEST_CODE"), function_name: Some("SNFORGE_TEST_CODE_FUNCTION") }!
+            [ERROR] There are no syscalls left in the program trace, but at least one unhandled call in trace file CallEntryPoint { class_hash: Some(ClassHash(0x117)), entry_point_type: External, entry_point_selector: EntryPointSelector(0x17340c6779204ea2a91c87d1c2226a3aebda65c64da3672a36893c4330ea27b), contract_address: ContractAddress(0x1724987234973219347210837402), call_type: Call, contract_name: Some("SNFORGE_TEST_CODE"), function_name: Some("SNFORGE_TEST_CODE_FUNCTION"), calldata_len: Some(0) }!
             
-            thread 'main' panicked at crates/cairo-profiler/src/trace_reader.rs:249:13:
+            thread 'main' panicked at crates/cairo-profiler/src/trace_reader.rs:254:13:
             Too many EntryPointCalls for triggers
             note: run with `RUST_BACKTRACE=1` environment variable to display a backtrace
             "#
@@ -988,7 +988,7 @@ fn tree_more_triggers_than_nested_calls() {
             "
             [ERROR] Found syscall CallContract in the program trace, that do not have corresponding calls in trace file!
             
-            thread 'main' panicked at crates/cairo-profiler/src/trace_reader.rs:198:17:
+            thread 'main' panicked at crates/cairo-profiler/src/trace_reader.rs:203:17:
             Too few EntryPointCalls for triggers
             note: run with `RUST_BACKTRACE=1` environment variable to display a backtrace
             "
@@ -1015,9 +1015,9 @@ fn tree_mismatched_syscall_with_entrypoint() {
         .failure()
         .stderr_eq(indoc!(
             r#"
-            [ERROR] Found syscall CallContract in the program trace, that do not corresponds to the next call from trace file CallEntryPoint { class_hash: Some(ClassHash(0x117)), entry_point_type: External, entry_point_selector: EntryPointSelector(0x17340c6779204ea2a91c87d1c2226a3aebda65c64da3672a36893c4330ea27b), contract_address: ContractAddress(0x1724987234973219347210837402), call_type: Call, contract_name: Some("SNFORGE_TEST_CODE"), function_name: Some("SNFORGE_TEST_CODE_FUNCTION") }!
+            [ERROR] Found syscall CallContract in the program trace, that do not corresponds to the next call from trace file CallEntryPoint { class_hash: Some(ClassHash(0x117)), entry_point_type: External, entry_point_selector: EntryPointSelector(0x17340c6779204ea2a91c87d1c2226a3aebda65c64da3672a36893c4330ea27b), contract_address: ContractAddress(0x1724987234973219347210837402), call_type: Call, contract_name: Some("SNFORGE_TEST_CODE"), function_name: Some("SNFORGE_TEST_CODE_FUNCTION"), calldata_len: Some(0) }!
             
-            thread 'main' panicked at crates/cairo-profiler/src/trace_reader.rs:239:17:
+            thread 'main' panicked at crates/cairo-profiler/src/trace_reader.rs:244:17:
             Trigger does not match entrypoint
             note: run with `RUST_BACKTRACE=1` environment variable to display a backtrace
             "#